newlib-SP1 + classic: replaced old sp1.h include for classic by newlib version
